Rename pasteDelegate to jsq_pasteDelegate
In the iOS 11.0 SDK Apple added UITextPasteConfigurationSupporting
which made pasteDelegate a property on UITextView which
JSQMessagesComposerTextView inherits. This class also defines a
pasteDelegate of type JSQMessagesComposerTextViewPasteDelegate, but
it ends up being an object confirming to the UITextPaste… protocol
that causes an unrecognized selector and crashes.

There are several other fixes available but most seem to agree that
renaming this delegate is the most appropriate solution.